+++ src/hooks/hermes/session-end.ts
@@ -1,0 +1,9 @@
+#!/usr/bin/env node
+import { runSessionEnd } from '../public/sessionEndPublic.js';
+import { isMemoryAnchorWorkspace } from './workspaceGuard.js';
+
+if (isMemoryAnchorWorkspace()) {
+  void runSessionEnd();
+} else {
+  process.exit(0);
+}